add package structure
