+++ pass-grant-cli/src/main/java/org/dataconservancy/pass/grant/cli/CoeusGrantLoaderCLI.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2018 Johns Hopkins University
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.dataconservancy.pass.grant.cli;
+
+import org.kohsuke.args4j.Argument;
+import org.kohsuke.args4j.CmdLineException;
+import org.kohsuke.args4j.CmdLineParser;
+import org.kohsuke.args4j.Option;
+
+import java.io.File;
+
+public class CoeusGrantLoaderCLI {
+
+    /**
+     * Arguments - just the property file containing the submission elements
+     */
+    @Argument(required = true, index = 0, metaVar = "[properties file]", usage = "properties file for the COEUS connection")
+    public static File propertiesFile = null;
+
+    /**
+     *
+     * General Options
+     */
+
+    /** Request for help/usage documentation */
+    @Option(name = "-h", aliases = { "-help", "--help" }, usage = "print help message")
+    public boolean help = false;
+
+    /** Requests the current version number of the cli application. */
+    @Option(name = "-v", aliases = { "-version", "--version" }, usage = "print version information")
+    public boolean version = false;
+
+}
