+++ pass-grant-data/pom.xml
@@ -47,7 +47,6 @@
         <dependency>
             <groupId>com.oracle</groupId>
             <artifactId>ojdbc8</artifactId>
-            <version>12.1.0</version>
         </dependency>
         <dependency>
             <groupId>junit</groupId>
