+++ pass-grant-model/src/main/java/org/dataconservancy/pass/grant/model/Person.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.dataconservancy.pass.grant.model;
 
 import java.util.List;
